Improvements over the Acquia pull files and the pantheon pull files logic

+++ src/Robo/Plugin/Commands/GetFilesCommand.php
@@ -18,9 +18,10 @@ class GetFilesCommand extends FireCommandBase {
    * @command local:get-files
    * @aliases get-files, files-get, getfiles, filesget, get_files, files_get, pull-files, pull_files, local:file:get, local:get:files
    * @usage fire local:files:get
-   * @option $no-overwrite If the value is TRUE, checks if the files exists and they are no downloaded again.
+   * @option $no-download Reuse your existing files copy in the reference folder and placing them in the files folder (Pantheon only).
    */
-  public function getFiles(ConsoleIO $io, $opts = ['no-overwrite' => FALSE]) {
+  public function getFiles(ConsoleIO $io, $opts = ['no-download' => FALSE]) {
+    $tasks = $this->collectionBuilder($io);
     $remotePlatform = Robo::config()->get('remote_platform');
     $remoteSiteName = Robo::config()->get('remote_sitename');
     $remoteEnv = Robo::config()->get('remote_canonical_env');
@@ -29,16 +30,27 @@ class GetFilesCommand extends FireCommandBase {
 
     // Create the folder.
     if (!file_exists($origFilesFolder)) {
-      $this->_mkdir($origFilesFolder);
+      $tasks->addTask($this->taskFilesystemStack()->mkdir($origFilesFolder));
+    }
+
+    // Creating files the folder if not exists.
+    if (!file_exists($destFilesFolder)) {
+      $this->taskFilesystemStack()->mkdir($destFilesFolder);
+    }
+    else {
+      // Removing all existing files from files folder.
+      if ($remotePlatform == 'pantheon') {
+        $tasks->addTask($this->taskCleanDir($destFilesFolder));
+      }
     }
 
     switch ($remotePlatform) {
       case 'acquia':
-        $tasks = $this->getFilesAcquia($io, $opts, $remotePlatform, $remoteSiteName, $remoteEnv, $origFilesFolder, $destFilesFolder);
+        $tasks = $this->getFilesAcquia($io, $tasks, $remoteSiteName, $remoteEnv);
         break;
       case 'pantheon':
       default:
-        $tasks = $this->getFilesPantheon($io, $opts, $remotePlatform, $remoteSiteName, $remoteEnv, $origFilesFolder, $destFilesFolder);
+        $tasks = $this->getFilesPantheon($io, $opts, $tasks, $remoteSiteName, $remoteEnv, $origFilesFolder, $destFilesFolder);
         break;
     }
 
@@ -48,59 +60,42 @@ class GetFilesCommand extends FireCommandBase {
   /**
    * Helper function to get files from Acquia.
    */
-  private function getFilesAcquia(ConsoleIO $io, array $opts, $remotePlatform, $remoteSiteName, $remoteEnv, $origFilesFolder, $destFilesFolder) {
-    $tasks = $this->collectionBuilder($io);
-
-    if ($this->getCliToolStatus('acli')) {
-      $uploadFolder = 'docroot';
-      if (!$opts['no-overwrite']) {
-        $cmd = "cd $origFilesFolder";
-        $cmd .= " && acli pull:files $remoteSiteName.$remoteEnv default";
-        $cmd .= " && tar czvf $uploadFolder.tar.gz $uploadFolder";
-        $this->taskExec($cmd)->run();
-      }
-
-      if (file_exists("$origFilesFolder/$uploadFolder.tar.gz")) {
-        if (!file_exists("$origFilesFolder/$uploadFolder")) {
-          $this->taskExec("cd $origFilesFolder && tar xzvf $uploadFolder.tar.gz")->run();
-        }
-
-        $tasks->addTask($this->taskExec("cp -a $origFilesFolder/$uploadFolder/sites/default/files/* $destFilesFolder"));
-        $tasks->addTask($this->taskExec("chmod -R 775 $origFilesFolder/$uploadFolder"));
-        $tasks->addTask($this->taskExec("rm -r $origFilesFolder/$uploadFolder"));
+  private function getFilesAcquia(ConsoleIO $io, $tasks, $remoteSiteName, $remoteEnv) {
+    /**
+     * Acquia CLI will automatically placed all files under docroot/sites/default/files.`
+     */
+    if (file_exists($this->getLocalEnvRoot() . '/docroot/sites/default/files')) {
+      if ($this->getCliToolStatus('acli')) {
+          $io->say('Syncing files from the ' . $remoteEnv . ' environment...');
+          $cmd = 'cd ' . $this->getLocalEnvRoot();
+          $cmd .= ' && acli pull:files ' . $remoteSiteName . '.' . $remoteEnv . ' default';
+          $tasks->addTask($this->taskExec($cmd));
       }
       else {
-        return "The folder '$uploadFolder' doesn't exist.";
+        return 'Acquia CLI is not installed, please install and configure it: https://docs.acquia.com/acquia-cli/install/';
       }
     }
     else {
-      return 'Acquia CLI is not installed, please install and configure it: https://docs.acquia.com/acquia-cli/install/';
+      return 'Your Local env doesnt have a valid acquia project structure, your drupal root should be the "docroot" folder.';
     }
-
     return $tasks;
   }
 
   /**
    * Helper function to get files from Pantheon.
    */
-  private function getFilesPantheon(ConsoleIO $io, array $opts, $remotePlatform, $remoteSiteName, $remoteEnv, $origFilesFolder, $destFilesFolder) {
-    $tasks = $this->collectionBuilder($io);
+  private function getFilesPantheon(ConsoleIO $io, array $opts, $tasks, $remoteSiteName, $remoteEnv, $origFilesFolder, $destFilesFolder) {
 
     if ($this->getCliToolStatus('terminus')) {
       $filesName = 'site-files.tar.gz';
-      if (!$opts['no-overwrite']) {
+      if (!$opts['no-download']) {
         $cmd = "wget `terminus backup:get $remoteSiteName.$remoteEnv --element=files` -O $origFilesFolder/$filesName";
-        $this->taskExec($cmd)->run();
+        $tasks->addTask($this->taskExec($cmd));
       }
-
-      if (file_exists("$origFilesFolder/$filesName")) {
-        $tasks->addTask($this->taskExec("cd $origFilesFolder && tar xzvf $filesName"));
-        $tasks->addTask($this->taskExec("cp -a $origFilesFolder/files_$remoteEnv/* $destFilesFolder"));
-        $tasks->addTask($this->taskExec("rm -r $origFilesFolder/files_$remoteEnv"));
-      }
-      else {
-        return "The file '$filesName' doesn't exist.";
-      }
+      $tasks->addTask($this->taskFilesystemStack()->mkdir($origFilesFolder . '/files_' . $remoteEnv));
+      $tasks->addTask($this->taskExec('tar xzvf ' . $origFilesFolder . '/' . $filesName . ' -C ' . $origFilesFolder . '/files_' . $remoteEnv));
+      $tasks->addTask($this->taskCopyDir([$origFilesFolder . '/files_' . $remoteEnv => $destFilesFolder]));
+      $tasks->addTask($this->taskDeleteDir($origFilesFolder . '/files_' . $remoteEnv));
     }
     else {
       return 'Terminus is not installed, please install and configure it: https://docs.pantheon.io/terminus/install';
